requirements.txt
langchain-ollama>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
